update backend
- api for transacction (deposit money to account, donation function, save transaction to history)
- make changes to model (***affect database***)
- add exception to prevent user from donating more than the money they have in the account

+++ DonationBackend/src/main/java/com/example/demo/controller/OrganizationController.java
@@ -1,7 +1,6 @@
 package com.example.demo.controller;
 
 import com.example.demo.exception.OrganizationNotFoundException;
-import com.example.demo.exception.UserNotFoundException;
 import com.example.demo.model.Organization;
 import com.example.demo.repository.OrganizationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,13 +25,13 @@ public class OrganizationController {
     }
 
     @GetMapping("/organization/{id}")
-    Organization getOrgById(@PathVariable Short id) {
+    Organization getOrgById(@PathVariable int id) {
         return organizationRepository.findById(id)
                 .orElseThrow(() -> new OrganizationNotFoundException(id));
     }
 
     @PutMapping("/organization/{id}")
-    Organization updateOrganization(@RequestBody Organization newOrganization, @PathVariable Short id) {
+    Organization updateOrganization(@RequestBody Organization newOrganization, @PathVariable int id) {
         return organizationRepository.findById(id)
                 .map(org -> {
                     org.setDescription(newOrganization.getDescription());
@@ -42,7 +41,7 @@ public class OrganizationController {
     }
 
     @DeleteMapping("/organization/{id}")
-    String deleteOrganization(@PathVariable Short id) {
+    String deleteOrganization(@PathVariable int id) {
         if (!organizationRepository.existsById(id)) {
             throw new OrganizationNotFoundException(id);
         }
