+++ DonationBackend/src/main/java/com/example/demo/exception/OrganizationNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.demo.exception;
 
 public class OrganizationNotFoundException extends RuntimeException {
-    public OrganizationNotFoundException(Short id) {
+    public OrganizationNotFoundException(int id) {
         super("Could not find the organization with id: " + id);
     }
 }
