+++ DonationBackend/src/main/java/com/example/demo/exception/ProgramNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.demo.exception;
 
 public class ProgramNotFoundException extends RuntimeException {
-    public ProgramNotFoundException(Short id) {
+    public ProgramNotFoundException(int id) {
         super("Could not find the program with id: " + id);
     }
 }
