+++ DonationBackend/src/main/java/com/example/demo/repository/OrganizationRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Organization;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Short> {
+public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
 }
