+++ DonationBackend/src/main/java/com/example/demo/repository/ProgramRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Program;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProgramRepository extends JpaRepository<Program, Short> {
+public interface ProgramRepository extends JpaRepository<Program, Integer> {
 }
